Add: Wide format support New-TypeFormatData (#1)
Adds the wide format to the new-TypeFormatData command so that wide displays can be created with it.

Wide settings are fairly simple so mostly fit in with the current settings. The auto size and column numbers are not part of a property setting so are configured with a new parameter ViewOptions.